LayoutTests/platform/glib/imported/w3c/web-platform-tests/content-security-policy/reporting-api/reporting-api-works-on-frame-ancestors.https.sub-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'report-to'.

CONSOLE MESSAGE: Refused to load https://not-web-platform.test:9443/content-security-policy/reporting-api/support/non-embeddable-frame.html?id=9ab0f605-6fad-4b75-8e90-7f133abf5e4f because it does not appear in the frame-ancestors directive of the Content Security Policy.


FAIL Violation report status OK. undefined is not an object (evaluating 'data[0]["body"]')
PASS No report to parent.
